Airplane-Game-in-Unity3d-using-Reinforcement-Learning
This Project is a convergence of AI and Video Games. This game is developed using Unity Game Engine along with aid from their Built in ML Agents Package. In this game you race along with other AI controlled airplanes.The goal is to fly through each checkpoint and be the first to reach the final checkpoint to finish first place.